Prepare for new carrier release. (#771)
* Checking in progress.

* Checking in progress.

* Undo hardcoded dev fictitious host.

* Small cleanup.

* Adding small delay.

* Make feather local safe.

* Update for compatibility with latest tierceron hat

* Add host finding logic.

* Adding env raw.

* Fixing problem with trcplgtool setup.

* Fixing incorrect sent code.

* Fix various auth and state transition issues.

* Implement interagent ctl messaging.

* Change flapping mode for inter agent communication.

* Fix missing import.

* Adding trcsh.exe to plugin build.

* Refactor to avoid package naming collision from client capauth.

---------
