+++ Makefile.tc
@@ -84,6 +84,9 @@ trcshellsha:
 	sha256sum trcvault/deploy/target/trcsh | cut -d' ' -f1 > trcvault/deploy/target/trcsh.sha256
 trcshell: trcshellbuild trcshellsha
 
+trcshellwin:
+	@GOPATH=$(GOPATH) GOBIN=$(GOBIN) GOOS=windows GOARCH=amd64 go build -tags "tc windows azrcr memonly" -o $(GOBIN)/trcsh.exe github.com/trimble-oss/tierceron/trcsh
+
 trckubectlbuild:
 	@GOPATH=$(GOPATH) GOBIN=$(GOBIN) go build  -o trcvault/deploy/target/kubectl -tags "memonly" github.com/trimble-oss/tierceron/trcsh
 trckubectlsha:
