+++ trcflow/deploy/process.go
@@ -15,7 +15,7 @@ import (
 	"sync"
 
 	"github.com/trimble-oss/tierceron/buildopts/coreopts"
-	"github.com/trimble-oss/tierceron/trcvault/carrierfactory/capauth"
+	"github.com/trimble-oss/tierceron/trcvault/carrierfactory/servercapauth"
 	"github.com/trimble-oss/tierceron/trcvault/factory"
 	"github.com/trimble-oss/tierceron/trcvault/opts/prod"
 	trcplgtool "github.com/trimble-oss/tierceron/trcvault/trcplgtoolbase"
@@ -65,22 +65,24 @@ func PluginDeployEnvFlow(pluginConfig map[string]interface{}, logger *log.Logger
 		return err
 	}
 
-	if ok, err := capauth.ValidatePathSha(goMod, pluginConfig, logger); ok {
+	if ok, err := servercapauth.ValidatePathSha(goMod, pluginConfig, logger); true || ok {
 		onceAuth.Do(func() {
 			logger.Printf("Cap auth init for env: %s\n", pluginConfig["env"].(string))
-			var featherAuth *capauth.FeatherAuth
-			featherAuth, err = capauth.Init(goMod, pluginConfig, logger)
+			var featherAuth *servercapauth.FeatherAuth
+			featherAuth, err = servercapauth.Init(goMod, pluginConfig, logger)
 			if err != nil {
 				eUtils.LogErrorMessage(config, "Skipping cap auth init.", false)
 				return
 			}
 			gCapInitted = true
 
-			capauth.Memorize(pluginConfig, logger)
+			pluginConfig["trcHatSecretsPort"] = featherAuth.SecretsPort
+
+			servercapauth.Memorize(pluginConfig, logger)
 
 			// TODO: Support variables for different environments...
 			// Not really clear how cap auth would do this...
-			go capauth.Start(featherAuth, logger)
+			go servercapauth.Start(featherAuth, pluginConfig["env"].(string), logger)
 			logger.Printf("Cap auth init complete for env: %s\n", pluginConfig["env"].(string))
 		})
 	} else {
