+++ trcvault/deploy/refreshcarriertoken.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+if [[ -z "${VAULT_ENV}" ]]; then
+echo "Enter vault environment being configured: "
+read VAULT_ENV
+fi
+
 if [[ -z "${AGENT_VAULT_ADDR}" ]]; then
 echo "Enter agent vault host base url including port: "
 read AGENT_VAULT_ADDR
@@ -10,7 +15,7 @@ read SECRET_VAULT_ADDR
 fi
 
 if [[ -z "${SECRET_VAULT_PLUGIN_TOKEN}" ]]; then
-echo "Enter organization vault plugin token for certification: "
+echo "Enter organization vault plugin token for certification(config_token_plugin$VAULT_ENV): "
 read SECRET_VAULT_PLUGIN_TOKEN
 fi
 
@@ -19,13 +24,8 @@ echo "Enter agent vault root token: "
 read VAULT_TOKEN
 fi
 
-if [[ -z "${VAULT_ENV}" ]]; then
-echo "Enter vault environment being configured: "
-read VAULT_ENV
-fi
-
 if [[ -z "${VAULT_ENV_TOKEN}" ]]; then
-echo "Enter agent vault *plugin* environment token with tightly confined write permissions: "
+echo "Enter agent vault *plugin* environment token with tightly confined write permissions(config_token_plugin$VAULT_ENV): "
 read VAULT_ENV_TOKEN
 fi
